connect the dots
Arrays loaded if the current user can be identified.
Message editor loads the contact form for the particular user and with the desired purpose.

Message sender introduced into the window for contact forms to access (given those contact forms want to use the ipfs service... contact forms are at liberty to send a message in any way the user recipients has programmed or asked to be programmed.)