Message sending clarity
Active user to editor
Notion of identity... Keeping an object with identity and for which user_iinfo is a field.
Identity object top level keeps a reference to the directory structure, to recover cids of asset files. And, the private key is kept in local storage (indexedDB) by attaching it to the identy object.