FIx the conf file a litte.
Set up the default path for an input template file.
The form_link is now an object with two fields, link and html. The client app may supply a local file path or it may supply a cid for a know template in order to HTML to store as the user's default contact template.