Make it difficult to overwrite current crypto pathways
They do not actually overwrite ever. But, there is the possibility of making it seem that way if a use uploads identity to another browser and then stores the identity (create my profile). So, first disable the profile create button if there is a current identity. And, check for the existence of the clear directory if the browser does not have an identity and so thinks it's OK to send creation commands. Inform people that they need to upload their identity when they switch to new browsers.